Use let instead of var
'let' creates separate instances of the pos variable in each iteration
(block scope),
which are part of the closure of httpGetAsync's callback function. 'var'
creates one function-scoped variable that changes in each iteration, so
the next iteration changes the value of the variable in the closure of
the previous iteration!